requirements.txt
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.0.0
streamlit>=1.0.0
langchain-google-genai>=0.0.6
